backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
anthropic==0.34.2
pdfplumber==0.11.4
httpx==0.27.2
python-dotenv==1.0.1
